python-flask-app/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
pymongo==4.3.3
bcrypt==3.2.0
PyJWT==2.3.0
requests==2.26.0
python-dotenv==0.19.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
Werkzeug==2.0.3